+++ zapit/+zapit/+gui/+main/@controller/dispFrame.m
@@ -1,7 +1,7 @@
 function dispFrame(obj,~,~)
     % This callback is run every time a frame has been acquired
     %
-    %  function zapit.gui.main.controller
+    %  function zapit.gui.main.controller()
     %
     % Purpose
     % This callback listens to the lastAcuireFrame property on the model